netshell/requirements.txt
# NetShell dependencies
pyyaml>=6.0
jsonschema>=4.0
jinja2>=3.0
